Improve performance
I noticed that, with larger grids, the process would significantly
slow down. My guess is that having multiple event listeners on every
single cell of the grid, while having a few thousands of cells, was
taking up too much memory, so I tried event delegation and added the
event listeners only to the parent element (the grid - this is also
a practical solution since the child elements are recreated every time
the grid size is changed, thus their listeners need to be re-added
every time, while the grid itself is never replaced).