Fix mobile support
Ok, having never used touch events before, I naively thought that 'touchmove'
would work exactly like 'mouseover' and return, as its target, the element
that was being touched whenever the touch moved. That is not the case:
the target of a 'touchmove' event is always the element that was touched
when the touch started. As far as I can tell, there's no direct way to get
the elements on the path of a touchmove.
So I wrote a function, invoked only by a touchmove event, that gets the
coordinates of the touch while moving on the screen and calculates the
corresponding element.
Also, the hover effect wasn't working properly on mobile, so I added a
media query to add the hover effect only if hover is supported.